fix: updated yazi y-function to not require quitting twice

+++ modules/home-manager/fish.nix
@@ -1,7 +1,7 @@
 { config, pkgs, lib, inputs, ... }:
 {
   programs.fish = {
-    
+
     enable = true;
 
     shellAliases = {
@@ -10,11 +10,20 @@
 
     functions.y = ''
       set tmp (mktemp -t "yazi-cwd.XXXXXX")
+
+      # Run yazi and immediately store the exit status
       yazi $argv --cwd-file="$tmp"
-      if set cwd (command cat -- "$tmp"); and [ -n "$cwd" ]; and [ "$cwd" != "$PWD" ]
-        builtin cd -- "$cwd"
+      set exit_code $status
+
+      # Clean up and possibly cd if the temp file exists and is valid
+      if test -e "$tmp"
+        if set cwd (command cat -- "$tmp"); and test -n "$cwd"; and test "$cwd" != "$PWD"
+          builtin cd -- "$cwd"
+        end
+        rm -f -- "$tmp"
       end
-      rm -f -- "$tmp"
+
+      return $exit_code
     '';
   };
 }
